+++ src/components/item-sidebars/ShowItemSidebar.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useRecoilState, useSetRecoilState } from 'recoil'
 import client from '../../api/client'
 import { itemsState } from '../../global-state/itemsState'
@@ -24,6 +24,10 @@ const ShowItemSidebar = () => {
     const [sidebarHistory, setSidebarHistory] = useRecoilState(
         sidebarHistoryState
     )
+
+    useEffect(() => {
+        console.log('Showitemsidebar mounted')
+    }, [])
 
     const addItem = async () => {
         try {
