+++ src/pages/items/ItemsPage.tsx
@@ -29,7 +29,7 @@ const containerVariants = {
     show: {
         opacity: 1,
         transition: {
-            staggerChildren: 0.1,
+            staggerChildren: 0.01,
         },
     },
 }
