+++ src/types/state/shoppingListTypes.ts
@@ -1,7 +1,7 @@
 export interface shopListInfoStateInterface {
     activeListId: number
     created_at?: string
-    name?: string
+    name: string
     status: string
     updated_at?: string
     user_id?: number
